feat(m1a): workspace/member/me HTTP 路由 + require_member/require_role 中间件 (LUM-1342)
- routes/workspaces.rs：9 条 member-visible 路由，路径/状态码与上游
  router.go + workspace.go/auth.go 对齐；axum 0.7 用 :id 语法（{id} 为字面量段）
- middleware/authn.rs：占位 session（X-Multica-Session / multica_session cookie；
  store 命中 → user_id，未命中 UUID 直通供本地联调），AuthUser/WorkspacePathId/
  WorkspaceContext 写入 request extensions；require_user/require_member/require_role
  （角色集合为运行时参数）；401/404/403 与上游 middleware 一致
- mount_slice_workspace_member 接线 .merge(workspaces::router(state))；
  router(state) 签名贯穿（axum from_fn_with_state 需要构建期 state）；
  替换 mount.rs 中 /api/workspaces 的 M0 占位（axum 同 path+method 重复注册 panic）
- 测试：mc-repos DB 测试修 FK（member 需真实 user）+ slug/邮箱幂等，9/9 ×2 遍通过；
  tests/smoke.rs 追加 workspace_member_http_e2e（401/201/409/200/404/403/leave/
  delete 全链路），经 [[test]] path 挂为 mc-http 可运行 target
- docs/05-M1-WORKSPACE-MEMBER.md：路由清单 + 上游行号对应 + 未覆盖项/差异
- M0 基线编译修复（缺依赖/缺 import/测试缺陷），与 B/C 切片逐字节同构，
  按 docs/09 §7.2 合并时任取一方

验证：cargo build --workspace / cargo test --workspace（含 PG e2e）/
cargo clippy --workspace --all-targets 0 error / rustfmt（本切片文件）。
Refs LUM-1342

+++ crates/mc-repos/src/user.rs
@@ -176,7 +176,7 @@ where
     }
 
     async fn list(&self, filter: UserFilter) -> Result<Vec<User>> {
-        let limit: i64 = filter.limit.unwrap_or(100).min(500);
+        let limit: i64 = filter.limit.unwrap_or(100).min(500) as i64;
         let after_id = filter.after_id;
         let rows = sqlx::query_as::<_, UserRow>(
             "SELECT id, name, email, avatar_url, email_verified_at, language, timezone, \
@@ -242,6 +242,7 @@ impl UserRepo {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::Repository;
 
     #[test]
     fn filter_default_shape() {
@@ -277,8 +278,8 @@ mod tests {
     #[ignore]
     #[tokio::test]
     async fn db_upsert_by_email_is_idempotent() {
-        let url = std::env::var("MULTICA_TEST_DATABASE_URL")
-            .expect("set MULTICA_TEST_DATABASE_URL");
+        let url =
+            std::env::var("MULTICA_TEST_DATABASE_URL").expect("set MULTICA_TEST_DATABASE_URL");
         let pool = mc_db::pool::Db::connect(&url, 4, 1).await.unwrap();
         let repo = UserRepo::new(pool);
 
@@ -308,8 +309,8 @@ mod tests {
     #[ignore]
     #[tokio::test]
     async fn db_get_by_email_returns_none_on_missing() {
-        let url = std::env::var("MULTICA_TEST_DATABASE_URL")
-            .expect("set MULTICA_TEST_DATABASE_URL");
+        let url =
+            std::env::var("MULTICA_TEST_DATABASE_URL").expect("set MULTICA_TEST_DATABASE_URL");
         let pool = mc_db::pool::Db::connect(&url, 4, 1).await.unwrap();
         let repo = UserRepo::new(pool);
         let none = repo
@@ -322,8 +323,8 @@ mod tests {
     #[ignore]
     #[tokio::test]
     async fn db_update_me_partial_fields() {
-        let url = std::env::var("MULTICA_TEST_DATABASE_URL")
-            .expect("set MULTICA_TEST_DATABASE_URL");
+        let url =
+            std::env::var("MULTICA_TEST_DATABASE_URL").expect("set MULTICA_TEST_DATABASE_URL");
         let pool = mc_db::pool::Db::connect(&url, 4, 1).await.unwrap();
         let repo = UserRepo::new(pool);
         let u = repo
